timeline interaction improvements
