+++ packages/app/app/studio/[organization]/clips/[stageId]/ClipContext.tsx
@@ -1,5 +1,13 @@
 'use client';
-import React, { createContext, useContext, useState, useRef } from 'react';
+import React, {
+  createContext,
+  useContext,
+  useState,
+  useRef,
+  useEffect,
+  useCallback,
+} from 'react';
+import useSearchParams from '@/lib/hooks/useSearchParams';
 import { IExtendedMarker } from '@/lib/types';
 
 type PlaybackStatus = {
@@ -36,6 +44,11 @@ type ClipContextType = {
   stageId: string;
   isCreatingClip: boolean;
   setIsCreatingClip: React.Dispatch<React.SetStateAction<boolean>>;
+  handleMouseDown: (marker: string, event: React.MouseEvent) => void;
+  handleMouseMove: (event: MouseEvent) => void;
+  handleMouseUp: () => void;
+  handleMarkerClick: (marker: IExtendedMarker) => void;
+  goToClickTime: (event: React.MouseEvent) => void;
 };
 
 const ClipContext = createContext<ClipContextType | null>(null);
@@ -49,6 +62,10 @@ export const ClipProvider = ({
   children: React.ReactNode;
   stageId: string;
 }) => {
+  const { handleTermChange, searchParams } = useSearchParams();
+
+  const start = searchParams.get('start');
+  const end = searchParams.get('end');
   const [playbackStatus, setPlaybackStatus] = useState<PlaybackStatus | null>(
     null
   );
@@ -60,6 +77,7 @@ export const ClipProvider = ({
     displayTime: 0,
     unix: 0,
   });
+
   const [isLoading, setIsLoading] = useState<boolean>(false);
   const [fragmentLoading, setFragmentLoading] = useState<boolean>(false);
   const videoRef = useRef<HTMLVideoElement>(null);
@@ -67,11 +85,191 @@ export const ClipProvider = ({
   const [selectedTooltip, setSelectedTooltip] = useState<string | null>(null);
   const [markers, setMarkers] = useState<IExtendedMarker[]>([]);
   const [isCreatingClip, setIsCreatingClip] = useState<boolean>(false);
+  const [initialMousePos, setInitialMousePos] = useState<number>(0);
+  const [initialMarkerPos, setInitialMarkerPos] = useState<number>(0);
+  const [updateTimeStart, setUpdateTimeStart] = useState<boolean>(false);
+  const [updateTimeEnd, setUpdateTimeEnd] = useState<boolean>(false);
+  const maxLength = videoRef.current?.duration || 0;
+  const pixelsPerSecond = 4;
+  const timelineWidth = maxLength * pixelsPerSecond;
+
+  useEffect(() => {
+    if (handleTermChange) {
+      console.log(
+        'startTime',
+        startTime.displayTime,
+        startTime.unix,
+        convertSecondsToUnix(startTime.displayTime)
+      );
+
+      handleTermChange([
+        { key: 'start', value: String(startTime.displayTime) },
+        { key: 'end', value: String(endTime.displayTime) },
+        { key: 'currentTime', value: String(videoRef.current?.currentTime) },
+      ]);
+    }
+  }, [
+    handleTermChange,
+    startTime.displayTime,
+    startTime.unix,
+    endTime.displayTime,
+  ]);
+
+  const convertSecondsToUnix = (seconds: number) => {
+    if (playbackStatus) {
+      // Get the current Unix time
+      const currentUnixTime = Date.now();
+
+      // Calculate the Unix timestamp for the given seconds
+      const unixTimestamp =
+        currentUnixTime + seconds * 1000 - playbackStatus.offset;
+
+      return unixTimestamp;
+    } else {
+    }
+  };
 
   const updateClipBounds = (start: number, end: number) => {
     setStartTime((prevState) => ({ ...prevState, displayTime: start }));
     setEndTime((prevState) => ({ ...prevState, displayTime: end }));
   };
+
+  useEffect(() => {
+    if (updateTimeStart && !fragmentLoading && playbackStatus) {
+      setStartTime({
+        unix: Date.now() - playbackStatus.offset,
+        displayTime: startTime.displayTime,
+      });
+      setUpdateTimeStart(false);
+    }
+    if (updateTimeEnd && !fragmentLoading && playbackStatus) {
+      setEndTime({
+        unix: Date.now() - playbackStatus.offset,
+        displayTime: endTime.displayTime,
+      });
+      setUpdateTimeEnd(false);
+    }
+  }, [
+    updateTimeStart,
+    updateTimeEnd,
+    videoRef,
+    startTime,
+    dragging,
+    playbackStatus,
+    setStartTime,
+    setEndTime,
+    endTime.displayTime,
+    fragmentLoading,
+  ]);
+
+  const handleMouseDown = (marker: string, event: React.MouseEvent) => {
+    setDragging(marker);
+    setSelectedTooltip(marker);
+
+    setInitialMousePos(event.clientX);
+    setInitialMarkerPos(
+      marker === 'start' ? startTime.displayTime : endTime.displayTime
+    );
+  };
+
+  const handleMouseMove = useCallback(
+    (event: MouseEvent) => {
+      if (dragging && videoRef.current && playbackStatus) {
+        const mouseDelta = event.clientX - initialMousePos;
+        const timeDelta = (mouseDelta / timelineWidth) * maxLength;
+        const newTime = Math.max(
+          0,
+          Math.min(maxLength, initialMarkerPos + timeDelta)
+        );
+
+        if (dragging === 'start') {
+          if (newTime >= 0 && newTime < endTime.displayTime) {
+            setStartTime({
+              unix: Date.now() - playbackStatus.offset,
+              displayTime: newTime,
+            });
+            setUpdateTimeStart(true);
+          }
+        } else if (dragging === 'end') {
+          if (
+            newTime > startTime.displayTime &&
+            newTime <= videoRef.current.duration
+          ) {
+            setEndTime({
+              unix: Date.now() - playbackStatus.offset,
+              displayTime: newTime,
+            });
+            setUpdateTimeEnd(true);
+          }
+        }
+
+        videoRef.current.currentTime = newTime;
+      }
+    },
+    [
+      dragging,
+      initialMousePos,
+      initialMarkerPos,
+      maxLength,
+      timelineWidth,
+      endTime.displayTime,
+      startTime.displayTime,
+      setStartTime,
+      setEndTime,
+      videoRef,
+      playbackStatus,
+    ]
+  );
+
+  const handleMouseUp = useCallback(() => {
+    setDragging(null);
+  }, []);
+
+  const handleMarkerClick = (marker: IExtendedMarker) => {
+    if (videoRef.current) {
+      videoRef.current.currentTime = marker.start;
+      setStartTime({
+        displayTime: marker.start,
+        unix: Date.now() - playbackStatus!.offset,
+      });
+
+      setEndTime({
+        displayTime: marker.end,
+        unix: Date.now() - playbackStatus!.offset,
+      });
+
+      videoRef.current.play();
+    }
+  };
+
+  const goToClickTime = (event: React.MouseEvent) => {
+    if (videoRef.current && playbackStatus) {
+      const timelineElement = event.currentTarget as HTMLElement;
+
+      const timelineRect = timelineElement.getBoundingClientRect();
+      const relativeClickX =
+        event.clientX - timelineRect.left + timelineElement.scrollLeft;
+      const clickTime = (relativeClickX / timelineWidth) * maxLength;
+      videoRef.current.currentTime = clickTime;
+    }
+  };
+
+  useEffect(() => {
+    const preventDefault = (e: Event) => e.preventDefault();
+
+    window.addEventListener('dragstart', preventDefault);
+    window.addEventListener('dragover', preventDefault);
+    window.addEventListener('mousemove', handleMouseMove);
+    window.addEventListener('mouseup', handleMouseUp);
+
+    return () => {
+      window.removeEventListener('mousemove', handleMouseMove);
+      window.removeEventListener('mouseup', handleMouseUp);
+      window.removeEventListener('dragstart', preventDefault);
+      window.removeEventListener('dragover', preventDefault);
+    };
+  }, [handleMouseMove, handleMouseUp, selectedTooltip, videoRef]);
+
   return (
     <ClipContext.Provider
       value={{
@@ -96,6 +294,11 @@ export const ClipProvider = ({
         stageId,
         isCreatingClip,
         setIsCreatingClip,
+        handleMouseDown,
+        handleMouseMove,
+        handleMouseUp,
+        handleMarkerClick,
+        goToClickTime,
       }}
     >
       {children}
